requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
ollama
python-dotenv
streamlit
sentence-transformers
torch
transformers
FlagEmbedding